fix: test errors in arithmetic_test

+++ prover/maths/common/smartvectorsext/arithmetic_basic.go
@@ -54,8 +54,10 @@ func InnerProduct(a, b smartvectors.SmartVector) fext.Element {
 	for i := 0; i < a.Len(); i++ {
 		var tmp fext.Element
 		a_, b_ := a.GetExt(i), b.GetExt(i)
+
 		tmp.Mul(&a_, &b_)
 		res.Add(&res, &tmp)
+
 	}
 
 	return res
@@ -264,13 +266,13 @@ func Sum(a smartvectors.SmartVector) (res fext.Element) {
 		for i := range v.window {
 			res.Add(&res, &v.window[i])
 		}
-		constTerm := fext.NewElement(uint32(v.totLen-len(v.window)), 0, 0, 0)
+		constTerm := fext.NewElement(int64(v.totLen-len(v.window)), 0, 0, 0)
 		constTerm.Mul(&constTerm, &v.paddingVal)
 		res.Add(&res, &constTerm)
 		return res
 
 	case *ConstantExt:
-		res := fext.NewElement(uint32(v.length), 0, 0, 0)
+		res := fext.NewElement(int64(v.length), 0, 0, 0)
 		res.Mul(&res, &v.val)
 		return res
